chore(tui): split TuiApp state methods
