+++ src/tui/state/mod.rs
@@ -1,6 +1,9 @@
 use bottom_pane_model::BottomPaneModel;
 mod bottom_pane_model;
+mod overlay_state;
+mod pending_interaction;
 mod persistence;
+mod runtime_snapshot;
 mod state_presets;
 #[cfg(test)]
 mod tests;
@@ -127,23 +130,9 @@ use rara_provider_catalog::{ModelCatalogProvider, fallback_models};
 use rara_state::state_db::StateDb;
 
 use super::queued_input::PendingFollowUpMessage;
-use crate::agent::{Agent, AgentExecutionMode, BashApprovalMode};
+use crate::agent::{AgentExecutionMode, BashApprovalMode};
 use crate::codex_model_catalog::{CodexModelOption, CodexReasoningOption};
 use crate::config::{ConfigManager, DEFAULT_CODEX_BASE_URL, OpenAiEndpointKind};
-use crate::tui::is_ssh_session;
-
-fn completed_interaction_role(kind: InteractionKind, source: Option<&str>) -> &'static str {
-    match kind {
-        InteractionKind::Approval => "Shell Approval Completed",
-        InteractionKind::PlanApproval => "Plan Decision",
-        InteractionKind::RequestInput => match source {
-            Some("plan_agent") => "Planning Question Answered",
-            Some("explore_agent") => "Exploration Question Answered",
-            Some(_) => "Sub-agent Question Answered",
-            None => "Question Answered",
-        },
-    }
-}
 
 pub fn input_requests_command_palette(input: &str) -> bool {
     let trimmed = input.trim_start();
@@ -213,59 +202,6 @@ fn startup_warning_for_config(config: &crate::config::RaraConfig) -> Option<Stri
 mod composer;
 
 impl TuiApp {
-    fn ensure_completed_interaction_entry(
-        &mut self,
-        kind: InteractionKind,
-        title: &str,
-        summary: &str,
-        source: Option<&str>,
-    ) {
-        let role = completed_interaction_role(kind, source).to_string();
-        let message = format!("{title}: {summary}");
-        let exists = self
-            .active_turn
-            .entries
-            .iter()
-            .chain(
-                self.committed_turns
-                    .iter()
-                    .flat_map(|turn| turn.entries.iter()),
-            )
-            .any(|entry| entry.role == role && entry.message == message);
-        if !exists {
-            self.active_turn
-                .entries
-                .push(TranscriptEntry::new(role, message));
-        }
-    }
-
-    fn plan_approval_interaction(&self) -> PendingInteractionSnapshot {
-        PendingInteractionSnapshot {
-            kind: InteractionKind::PlanApproval,
-            title: "Plan Ready".to_string(),
-            summary: self
-                .snapshot
-                .plan_explanation
-                .clone()
-                .unwrap_or_else(|| "Review the proposed plan before implementation.".to_string()),
-            options: Vec::new(),
-            note: None,
-            approval: None,
-            source: None,
-        }
-    }
-
-    fn set_plan_approval_interaction(&mut self, pending: bool) {
-        self.snapshot
-            .pending_interactions
-            .retain(|item| item.kind != InteractionKind::PlanApproval);
-        if pending {
-            self.snapshot
-                .pending_interactions
-                .push(self.plan_approval_interaction());
-        }
-    }
-
     pub fn new(cm: ConfigManager) -> anyhow::Result<Self> {
         let mut cfg = cm.load()?;
         cfg.apply_provider_environment_defaults();
@@ -1500,587 +1436,6 @@ impl TuiApp {
             });
         self.config.set_reasoning_effort(selected);
     }
-
-    pub fn sync_snapshot(&mut self, agent: &Agent) {
-        let runtime_context = agent.shared_runtime_context();
-        let existing_pending_approval_id = self
-            .pending_command_approval()
-            .and_then(|item| item.approval.as_ref())
-            .map(|approval| approval.tool_use_id.clone());
-        let existing_plan_completion = self
-            .completed_interaction(InteractionKind::PlanApproval)
-            .cloned();
-        let existing_pending_plan_approval = self.pending_plan_approval_interaction().cloned();
-        let existing_local_request_completion = self
-            .snapshot
-            .completed_interactions
-            .iter()
-            .find(|item| {
-                item.kind == InteractionKind::RequestInput && item.source.as_deref().is_some()
-            })
-            .cloned();
-        let existing_local_request_inputs = self
-            .snapshot
-            .pending_interactions
-            .iter()
-            .filter(|item| {
-                item.kind == InteractionKind::RequestInput && item.source.as_deref().is_some()
-            })
-            .cloned()
-            .collect::<Vec<_>>();
-        let mut pending_interactions = Vec::new();
-        if let Some(question) = agent.pending_user_input.as_ref() {
-            pending_interactions.push(PendingInteractionSnapshot {
-                kind: InteractionKind::RequestInput,
-                title: question.question.clone(),
-                summary: question.note.clone().unwrap_or_default(),
-                options: question.options.clone(),
-                note: question.note.clone(),
-                approval: None,
-                source: None,
-            });
-        }
-        if agent.pending_user_input.is_none() {
-            pending_interactions.extend(existing_local_request_inputs);
-        }
-        if let Some(item) = existing_pending_plan_approval {
-            pending_interactions.push(item);
-        }
-        if let Some(pending) = agent.pending_approval.as_ref() {
-            pending_interactions.push(PendingInteractionSnapshot {
-                kind: InteractionKind::Approval,
-                title: "Pending Approval".to_string(),
-                summary: pending.request.summary(),
-                options: Vec::new(),
-                note: None,
-                approval: Some(PendingApprovalSnapshot {
-                    tool_use_id: pending.tool_use_id.clone(),
-                    command: pending.request.summary(),
-                    allow_net: self
-                        .sandbox_network_access
-                        .load(std::sync::atomic::Ordering::Relaxed)
-                        || pending.request.allow_net,
-                    payload: pending.request.clone(),
-                }),
-                source: None,
-            });
-        }
-        let current_pending_approval_id = agent
-            .pending_approval
-            .as_ref()
-            .map(|pending| pending.tool_use_id.clone());
-        if current_pending_approval_id != existing_pending_approval_id {
-            self.approval_picker_idx = 0;
-        }
-        let mut completed_interactions = Vec::new();
-        if let Some(item) = agent.completed_user_input.as_ref() {
-            completed_interactions.push(CompletedInteractionSnapshot {
-                kind: InteractionKind::RequestInput,
-                title: item.title.clone(),
-                summary: item.summary.clone(),
-                source: None,
-            });
-        }
-        if let Some(item) = agent.completed_approval.as_ref() {
-            completed_interactions.push(CompletedInteractionSnapshot {
-                kind: InteractionKind::Approval,
-                title: item.title.clone(),
-                summary: item.summary.clone(),
-                source: None,
-            });
-        }
-        if let Some(item) = existing_local_request_completion {
-            completed_interactions.push(item);
-        }
-        if let Some(item) = existing_plan_completion {
-            completed_interactions.push(item);
-        }
-        for interaction in completed_interactions.iter() {
-            self.ensure_completed_interaction_entry(
-                interaction.kind,
-                interaction.title.as_str(),
-                interaction.summary.as_str(),
-                interaction.source.as_deref(),
-            );
-        }
-        let ext_counts = discover_extension_counts(&runtime_context.cwd);
-        let runtime_hook_count = self
-            .hook_runtime
-            .as_ref()
-            .map(|runtime| runtime.hook_count())
-            .unwrap_or(0);
-        self.snapshot = RuntimeSnapshot {
-            cwd: runtime_context.cwd,
-            branch: runtime_context.branch,
-            session_id: runtime_context.session_id,
-            history_len: runtime_context.history_len,
-            total_input_tokens: runtime_context.total_input_tokens,
-            total_output_tokens: runtime_context.total_output_tokens,
-            total_cache_hit_tokens: runtime_context.total_cache_hit_tokens,
-            total_cache_miss_tokens: runtime_context.total_cache_miss_tokens,
-            context_window_tokens: runtime_context.budget.context_window_tokens,
-            compact_threshold_tokens: runtime_context.budget.compact_threshold_tokens,
-            reserved_output_tokens: runtime_context.budget.reserved_output_tokens,
-            stable_instructions_budget: runtime_context.budget.stable_instructions_budget,
-            workspace_prompt_budget: runtime_context.budget.workspace_prompt_budget,
-            active_turn_budget: runtime_context.budget.active_turn_budget,
-            compacted_history_budget: runtime_context.budget.compacted_history_budget,
-            retrieved_memory_budget: runtime_context.budget.retrieved_memory_budget,
-            remaining_input_budget: runtime_context.budget.remaining_input_budget,
-            estimated_history_tokens: runtime_context.compaction.estimated_history_tokens,
-            compaction_count: runtime_context.compaction.compaction_count,
-            last_compaction_before_tokens: runtime_context.compaction.last_compaction_before_tokens,
-            last_compaction_after_tokens: runtime_context.compaction.last_compaction_after_tokens,
-            last_compaction_recent_files: runtime_context.compaction.last_compaction_recent_files,
-            last_compaction_boundary_version: runtime_context
-                .compaction
-                .last_compaction_boundary_version,
-            last_compaction_boundary_before_tokens: runtime_context
-                .compaction
-                .last_compaction_boundary_before_tokens,
-            last_compaction_boundary_recent_file_count: runtime_context
-                .compaction
-                .last_compaction_boundary_recent_file_count,
-            compaction_source_entries: runtime_context.compaction.source_entries,
-            plan_steps: runtime_context.plan.steps,
-            plan_explanation: runtime_context.plan.explanation,
-            pending_interactions,
-            completed_interactions,
-            todo_artifact_path: if agent.todo_state.is_some() {
-                Some(
-                    agent
-                        .session_manager
-                        .todo_file_path(&agent.session_id)
-                        .display()
-                        .to_string(),
-                )
-            } else {
-                None
-            },
-            todo: runtime_context.todo,
-            prompt_base_kind: runtime_context.prompt.base_prompt_kind,
-            prompt_section_keys: runtime_context.prompt.section_keys,
-            prompt_source_entries: runtime_context.prompt.source_entries,
-            prompt_source_status_lines: runtime_context.prompt.source_status_lines,
-            prompt_append_system_prompt: runtime_context.prompt.append_system_prompt,
-            prompt_warnings: runtime_context.prompt.warnings,
-            retrieval_source_entries: runtime_context.retrieval.entries,
-            retrieval_orchestration: runtime_context.retrieval.orchestration,
-            memory_selection: runtime_context.retrieval.memory_selection,
-            context_observability: runtime_context.observability,
-            assembly_entries: runtime_context.assembly.entries,
-            // ── Extensions ──────────────────────────────────────
-            extension_skill_count: agent.prompt_config().available_skills.len(),
-            extension_skill_scopes: {
-                let mut scopes: Vec<String> = agent
-                    .prompt_config()
-                    .available_skills
-                    .iter()
-                    .map(|s| s.scope.clone())
-                    .collect::<std::collections::BTreeSet<_>>()
-                    .into_iter()
-                    .collect();
-                scopes.sort();
-                scopes
-            },
-            extension_hook_count: ext_counts.0.max(runtime_hook_count),
-            extension_agent_count: ext_counts.1,
-            extension_agent_status_lines: ext_counts.2,
-        };
-        self.agent_execution_mode = agent.execution_mode;
-        self.bash_approval_mode = agent.bash_approval_mode;
-        self.populate_skill_picker_entries(agent);
-        self.persist_runtime_state();
-    }
-
-    pub fn populate_skill_picker_entries(&mut self, agent: &Agent) {
-        self.skill_picker_entries = agent
-            .prompt_config()
-            .available_skills
-            .iter()
-            .map(|s| SkillPickerEntry {
-                name: s.name.clone(),
-                title: s.title.clone().unwrap_or_else(|| s.name.clone()),
-                scope: s.scope.clone(),
-                enabled: !s.disable_model_invocation,
-                disable_model_invocation: s.disable_model_invocation,
-            })
-            .collect();
-        self.skill_picker_entries
-            .sort_by(|a, b| a.name.cmp(&b.name));
-    }
-
-    pub fn open_overlay(&mut self, overlay: Overlay) {
-        if matches!(overlay, Overlay::CommandPalette | Overlay::ModelSearch) {
-            self.command_palette_idx = 0;
-        }
-        if matches!(overlay, Overlay::ListPicker(ListPickerKind::Provider)) {
-            self.provider_picker_idx = selected_provider_family_idx_for_config(&self.config);
-            self.refresh_provider_connection_status();
-        }
-        if matches!(overlay, Overlay::ListPicker(ListPickerKind::Resume)) {
-            self.refresh_recent_threads_for_resume_picker();
-        }
-        if matches!(overlay, Overlay::ListPicker(ListPickerKind::UnifiedModel)) {
-            self.refresh_provider_connection_status();
-            self.model_picker_idx = self.selected_unified_preset_idx();
-            self.sync_reasoning_effort_picker();
-        }
-        if matches!(overlay, Overlay::ListPicker(ListPickerKind::Model)) {
-            let selected_family = self.selected_provider_family();
-            if matches!(selected_family, ProviderFamily::OpenAiCompatible) {
-                if !matches!(
-                    PROVIDER_FAMILIES
-                        .get(selected_provider_family_idx_for_config(&self.config))
-                        .map(|(family, _, _)| *family),
-                    Some(ProviderFamily::OpenAiCompatible)
-                ) {
-                    self.config.set_provider("openai-compatible");
-                }
-                if self.config.active_openai_profile_kind() == Some(OpenAiEndpointKind::Deepseek) {
-                    self.config.select_openai_profile(
-                        OpenAiEndpointKind::Custom.default_profile_id(),
-                        OpenAiEndpointKind::Custom.label(),
-                        OpenAiEndpointKind::Custom,
-                    );
-                }
-                self.model_picker_idx = 0;
-            } else if let Some(provider) = self.single_provider_for_selected_family() {
-                self.config.set_provider(provider.to_string());
-                self.model_picker_idx = self.selected_preset_idx();
-            }
-            self.sync_reasoning_effort_picker();
-        }
-        if matches!(
-            overlay,
-            Overlay::ListPicker(ListPickerKind::OpenAiEndpointKind)
-        ) {
-            self.openai_endpoint_kind_picker_idx = self
-                .selected_openai_profile_kind()
-                .and_then(|kind| {
-                    openai_profile_setup_kinds()
-                        .iter()
-                        .position(|candidate| *candidate == kind)
-                })
-                .unwrap_or(0);
-        }
-        if matches!(overlay, Overlay::ListPicker(ListPickerKind::OpenAiProfile)) {
-            self.sync_openai_profile_picker();
-        }
-        if matches!(overlay, Overlay::BaseUrlEditor) {
-            let provider_family = self.selected_provider_family();
-            self.base_url_input = self.config.base_url.clone().unwrap_or_else(|| {
-                if matches!(provider_family, ProviderFamily::OpenAiCompatible) {
-                    self.config
-                        .active_openai_profile_kind()
-                        .unwrap_or(OpenAiEndpointKind::Custom)
-                        .default_base_url()
-                        .to_string()
-                } else {
-                    "http://localhost:11434".to_string()
-                }
-            });
-            self.base_url_cursor_offset = None;
-        }
-        if matches!(overlay, Overlay::ApiKeyEditor) {
-            self.api_key_input.clear();
-            self.api_key_cursor_offset = None;
-        }
-        if matches!(overlay, Overlay::ModelNameEditor) {
-            self.model_name_input = self.config.model.clone().unwrap_or_else(|| {
-                self.selected_model_preset()
-                    .map(|(_, _, default_model)| default_model.to_string())
-                    .unwrap_or_default()
-            });
-            self.model_name_cursor_offset = None;
-        }
-        if matches!(overlay, Overlay::OpenAiProfileLabelEditor) {
-            let kind = self
-                .openai_profile_label_kind
-                .or_else(|| self.selected_openai_profile_kind())
-                .unwrap_or(OpenAiEndpointKind::Custom);
-            self.openai_profile_label_input = format!("{} profile", kind.label());
-            self.openai_profile_label_cursor_offset = None;
-        }
-        if matches!(overlay, Overlay::ListPicker(ListPickerKind::AuthMode)) {
-            self.auth_mode_idx = if is_ssh_session() { 1 } else { 0 };
-        }
-        if matches!(
-            overlay,
-            Overlay::ListPicker(ListPickerKind::ReasoningEffort)
-        ) {
-            self.sync_reasoning_effort_picker();
-        }
-        if matches!(overlay, Overlay::SkillsPicker) {
-            self.skill_picker_idx = 0;
-        }
-        if matches!(overlay, Overlay::Context) {
-            // Auto-hide the command palette when opening a full-screen modal
-            // so Esc dismisses only the modal, not the stale palette underneath.
-            if !matches!(overlay, Overlay::CommandPalette | Overlay::ModelSearch)
-                && matches!(self.overlay, Some(Overlay::CommandPalette))
-            {
-                self.hide_overlay();
-            }
-            self.context_scroll = 0;
-        }
-        self.overlay_stack.push(overlay);
-        self.overlay = Some(overlay);
-    }
-
-    /// Pop the top overlay from the stack without user-visible side
-    /// effects (preserves input, does not cancel setups).  Used when
-    /// the overlay becomes irrelevant due to an input change rather
-    /// than an explicit user action.
-    fn hide_overlay(&mut self) {
-        self.overlay_stack.pop();
-        self.overlay = self.overlay_stack.last().copied();
-        self.command_palette_idx = 0;
-    }
-
-    /// Keep the command-palette overlay in sync with the current
-    /// input.  Opens the palette when the user starts typing `/`,
-    /// and auto-hides it when the input contains a space (i.e. the
-    /// command has arguments and the user intends to submit).
-    pub fn sync_command_palette_with_input(&mut self) {
-        let should_show = input_requests_command_palette(self.bottom_pane.input.as_str());
-        match (should_show, &self.overlay) {
-            (true, None) => self.open_overlay(Overlay::CommandPalette),
-            (false, Some(Overlay::CommandPalette)) => self.hide_overlay(),
-            _ => {}
-        }
-    }
-
-    pub fn set_agent_execution_mode(&mut self, mode: AgentExecutionMode) {
-        self.agent_execution_mode = mode;
-    }
-
-    pub fn agent_execution_mode_label(&self) -> &'static str {
-        match self.agent_execution_mode {
-            AgentExecutionMode::Execute => "execute",
-            AgentExecutionMode::Plan => "plan",
-            AgentExecutionMode::Review => "review",
-        }
-    }
-
-    pub fn permission_mode_label(&self) -> &'static str {
-        self.permission_mode.label()
-    }
-
-    pub fn bash_approval_mode_label(&self) -> &'static str {
-        match self.bash_approval_mode {
-            BashApprovalMode::Always => "always",
-            BashApprovalMode::Once => "once",
-            BashApprovalMode::Suggestion => "suggestion",
-        }
-    }
-
-    pub fn pending_question_option_label(&self, index: usize) -> Option<String> {
-        self.pending_request_input()
-            .and_then(|interaction| interaction.options.get(index))
-            .map(|(label, _)| label.clone())
-    }
-
-    pub fn has_pending_approval(&self) -> bool {
-        self.pending_command_approval().is_some()
-    }
-
-    pub fn has_pending_plan_approval(&self) -> bool {
-        self.pending_plan_approval_interaction().is_some()
-    }
-
-    pub fn active_pending_interaction(&self) -> Option<ActivePendingInteraction<'_>> {
-        if let Some(snapshot) = self.pending_plan_approval_interaction() {
-            return Some(ActivePendingInteraction {
-                kind: ActivePendingInteractionKind::PlanApproval,
-                _snapshot: snapshot,
-            });
-        }
-        if let Some(snapshot) = self.pending_command_approval() {
-            return Some(ActivePendingInteraction {
-                kind: ActivePendingInteractionKind::ShellApproval,
-                _snapshot: snapshot,
-            });
-        }
-        if let Some(snapshot) = self.pending_request_input() {
-            let kind = match snapshot.source.as_deref() {
-                Some("plan_agent") => ActivePendingInteractionKind::PlanningQuestion,
-                Some("explore_agent") => ActivePendingInteractionKind::ExplorationQuestion,
-                Some(_) => ActivePendingInteractionKind::SubAgentQuestion,
-                None => ActivePendingInteractionKind::RequestInput,
-            };
-            return Some(ActivePendingInteraction {
-                kind,
-                _snapshot: snapshot,
-            });
-        }
-        None
-    }
-
-    pub fn active_pending_option_count(&self) -> usize {
-        let Some(pending) = self.active_pending_interaction() else {
-            return 0;
-        };
-        match pending.kind {
-            ActivePendingInteractionKind::PlanApproval => 2,
-            ActivePendingInteractionKind::ShellApproval => 4,
-            ActivePendingInteractionKind::PlanningQuestion
-            | ActivePendingInteractionKind::ExplorationQuestion
-            | ActivePendingInteractionKind::SubAgentQuestion
-            | ActivePendingInteractionKind::RequestInput => self
-                .pending_request_input()
-                .map(|interaction| interaction.options.len().min(3))
-                .unwrap_or(0),
-        }
-    }
-
-    pub fn set_pending_plan_approval(&mut self, pending: bool) {
-        self.set_plan_approval_interaction(pending);
-        self.persist_runtime_state();
-    }
-
-    pub fn pending_request_input(&self) -> Option<&PendingInteractionSnapshot> {
-        self.snapshot
-            .pending_interactions
-            .iter()
-            .find(|item| item.kind == InteractionKind::RequestInput)
-    }
-
-    pub fn has_local_pending_request_input(&self) -> bool {
-        self.pending_request_input()
-            .and_then(|item| item.source.as_deref())
-            .is_some()
-    }
-
-    pub fn pending_command_approval(&self) -> Option<&PendingInteractionSnapshot> {
-        self.snapshot
-            .pending_interactions
-            .iter()
-            .find(|item| item.kind == InteractionKind::Approval)
-    }
-
-    pub fn clear_pending_command_approval(&mut self) {
-        self.snapshot
-            .pending_interactions
-            .retain(|item| item.kind != InteractionKind::Approval);
-        self.persist_runtime_state();
-    }
-
-    pub fn pending_plan_approval_interaction(&self) -> Option<&PendingInteractionSnapshot> {
-        self.snapshot
-            .pending_interactions
-            .iter()
-            .find(|item| item.kind == InteractionKind::PlanApproval)
-    }
-
-    pub fn completed_interaction(
-        &self,
-        kind: InteractionKind,
-    ) -> Option<&CompletedInteractionSnapshot> {
-        self.snapshot
-            .completed_interactions
-            .iter()
-            .find(|item| item.kind == kind)
-    }
-
-    pub fn record_completed_interaction(
-        &mut self,
-        kind: InteractionKind,
-        title: impl Into<String>,
-        summary: impl Into<String>,
-        source: Option<String>,
-    ) {
-        let title = title.into();
-        let summary = summary.into();
-        self.snapshot
-            .completed_interactions
-            .retain(|item| item.kind != kind);
-        self.snapshot
-            .completed_interactions
-            .push(CompletedInteractionSnapshot {
-                kind,
-                title: title.clone(),
-                summary: summary.clone(),
-                source: source.clone(),
-            });
-        self.ensure_completed_interaction_entry(
-            kind,
-            title.as_str(),
-            summary.as_str(),
-            source.as_deref(),
-        );
-        self.persist_runtime_state();
-    }
-
-    pub fn record_local_request_input(
-        &mut self,
-        source: impl Into<String>,
-        title: impl Into<String>,
-        options: Vec<(String, String)>,
-        note: Option<String>,
-    ) {
-        self.snapshot.pending_interactions.retain(|item| {
-            !(item.kind == InteractionKind::RequestInput && item.source.as_deref().is_some())
-        });
-        let title = title.into();
-        self.snapshot
-            .pending_interactions
-            .push(PendingInteractionSnapshot {
-                kind: InteractionKind::RequestInput,
-                title: title.clone(),
-                summary: note.clone().unwrap_or_default(),
-                options,
-                note,
-                approval: None,
-                source: Some(source.into()),
-            });
-        self.bottom_pane.notice = Some(title.clone());
-        self.persist_runtime_state();
-    }
-
-    pub fn clear_local_request_input(&mut self) {
-        self.snapshot.pending_interactions.retain(|item| {
-            !(item.kind == InteractionKind::RequestInput && item.source.as_deref().is_some())
-        });
-        self.persist_runtime_state();
-    }
-
-    pub fn dismiss_overlay(&mut self) {
-        if matches!(
-            self.overlay,
-            Some(Overlay::BaseUrlEditor | Overlay::ApiKeyEditor | Overlay::ModelNameEditor)
-        ) {
-            self.cancel_openai_profile_setup();
-        }
-
-        // When dismissing the command palette, clear the `/` input so
-        // sync_command_palette_with_input won't immediately re-open it.
-        if matches!(
-            self.overlay,
-            Some(Overlay::CommandPalette | Overlay::ModelSearch)
-        ) {
-            self.bottom_pane.input.clear();
-            self.command_palette_idx = 0;
-        }
-
-        self.hide_overlay();
-        self.overlay = self.overlay_stack.last().copied();
-    }
-}
-
-fn discover_extension_counts(cwd: &str) -> (usize, usize, Vec<String>) {
-    let root = std::path::Path::new(cwd);
-    let mut hr = crate::hooks::HookRegistry::new();
-    let mut ar = crate::agents_ext::AgentRegistry::new();
-    hr.discover_repo_hooks(root);
-    ar.discover_repo_agents(root);
-    let agent_count = ar.agents.len();
-    let agent_status_lines = if agent_count == 0 {
-        Vec::new()
-    } else {
-        ar.status_lines()
-    };
-    (hr.hooks.len(), agent_count, agent_status_lines)
 }
 
 mod helpers;
